+++ cmd/libs/go2idl/client-gen/test_apis/testgroup/install/install.go
@@ -85,7 +85,7 @@ func newRESTMapper(externalVersions []schema.GroupVersion) meta.RESTMapper {
 
 	ignoredKinds := sets.NewString()
 
-	return api.NewDefaultRESTMapperFromScheme(externalVersions, interfacesFor, importPrefix, ignoredKinds, rootScoped, api.Scheme)
+	return meta.NewDefaultRESTMapperFromScheme(externalVersions, interfacesFor, importPrefix, ignoredKinds, rootScoped, api.Scheme)
 }
 
 // InterfacesFor returns the default Codec and ResourceVersioner for a given version
