+++ pkg/apis/authorization/install/BUILD
@@ -13,9 +13,9 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/api:go_default_library",
-        "//pkg/apimachinery/announced:go_default_library",
         "//pkg/apis/authorization:go_default_library",
         "//pkg/apis/authorization/v1beta1:go_default_library",
+        "//vendor:k8s.io/apimachinery/pkg/apimachinery/announced",
         "//vendor:k8s.io/apimachinery/pkg/apimachinery/registered",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
